% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{Ajuste_rapido}
\alias{Ajuste_rapido}
\title{Realiza analisis de manera directa}
\usage{
Ajuste_rapido(
  datos,
  frecuencia = NULL,
  inicio = NULL,
  init_ = TRUE,
  msg = FALSE,
  ...
)
}
\arguments{
\item{datos}{Datos para el analisis de serie de tiempo suport (data.frame and ts class)}

\item{frecuencia}{Frecuencia de la serie de tiempo, sirve para
reescribir la frecuencia cuando datos es un objeto ts}

\item{inicio}{Inicio de la serie de tiempo, igual que frecuencia
sobreescribe valores de objetos ts}

\item{init_}{(True or False) validar el parametro datos}

\item{...}{Not work}
}
\value{
La salida no es como tal un objeto, si no una serie de impresiones de varios
  analisis. La siguiente lista detalla alguno de ellos:
  \itemize{\item{Plost}{  Arroja una lista de plots que ayudan a ver el comportamiento de la serie y como ciertos ajustes se aproximan mejor a ella}}
  \itemize{\item{Resumenes}{  Arroja ciertos resumenes de ciertos ajustes o pruebas que se hacen}}
  \itemize{\item{Modelo}{  Modelo con el menor MSE(Error cuadratico medio)}}
}
\description{
Esta función considera que las recomendaciones de la paqueteria seran tomandas como los valores a usar
}
\details{
Todo los detalles son los mismo que en \code{\link{init}}
}
\examples{
 base=data.frame(tiempo=seq(Sys.Date(),by="days",length=20),valores=1:20*3+runif(1))
 Ajuste_rapido(datos=base,frecuencia=4,inicio=2010)

}
