% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{checar_datos}
\alias{checar_datos}
\title{Checa los datos en rutina TSR}
\usage{
checar_datos(datos, frecuencia, inicio)
}
\arguments{
\item{datos}{Data.Frame de 2 columnas, fecha y valores respectivamente. (se hace tratamiento de fechas con \code{\link{tratamiento.fechas.TRS}})}

\item{frecuencia}{Frecuencia de la serie de tiempo}

\item{inicio}{Inicio de la serie de tiempo}
}
\value{
Una lista \code{\link{list}} que contiene dos elementos, la base de datos tratada y un objeto TimeSeries
}
\description{
Imprime un head de los datos y te pregunta si todo esta bien.
}
\details{
Ideal para usarse en las rutinas de TSRutina que deben preguntar si se interpretaron
bien los datos.
}
\examples{
base=data.frame(tiempo=seq(Sys.Date(),by="days",length=20),valores=1:20*3+runif(1))
checar_datos(datos=base,frecuencia=4,inicio=2010)

}
