% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{conditional.tsrutina}
\alias{conditional.tsrutina}
\title{Primer filtro de las funciones en TSRutina}
\usage{
conditional.tsrutina(datos)
}
\arguments{
\item{datos}{objeto a verificar}
}
\value{
NULL or stop() event
}
\description{
Verifica si un data.frame tiene dos columnas y si la segunda de estas
es de caracter numerico, si es serie de tiempo no hace nada, retorna NULL.
}
\examples{

conditional.tsrutina(datos)

}
