% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ajuste_prophet.R
\encoding{UTF-8}
\name{data_to_prophet}
\alias{data_to_prophet}
\title{Data to prophet format}
\usage{
data_to_prophet(datos, ..., verbose = T)
}
\arguments{
\item{datos}{Datos para el análisis de serie de tiempo solo Data.frame}
}
\value{
Dataframe con formato para prophet
}
\description{
Agrega el formato para hacer modelado prophet
nombres de variable ds and y
}
\examples{

datos <- read.csv("https://raw.githubusercontent.com/crissthiandi/datos/master/Series_tiempo/sunspot_month_dataframe.csv")

data_to_prophet(datos)
}
