% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{dev.TRS}
\alias{dev.TRS}
\title{Depurador de dispositivos graficos}
\usage{
dev.TRS()
}
\value{
NULL return
}
\description{
Depura el espacio de dispositivos graficos atravez de dev.off() y dev.new(). Checando dev.list().
}
\examples{
dev.TRS()
#Se limpia los datos, consulte dev.list()
}
