% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{hacer_reporte}
\alias{hacer_reporte}
\title{De lista a reporte PDF}
\usage{
hacer_reporte(datos)
}
\arguments{
\item{datos}{Time Series a analizar}
}
\value{
PDF report
}
\description{
Haz un reporte de tus objetos lista para poder analizar en formato de lectura tu serie de tiempo
}
\details{
En desarrollo probablemente en uso hasta la version 3.1< de la tabby verse
}
\examples{

a=TSRutina::Ajuste_ARIMA_rapido(AirPassengers,pausa_off = 0)
hacer_reporte(a)

}
