% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ajuste_prophet.R
\encoding{UTF-8}
\name{holydays_to_prophet}
\alias{holydays_to_prophet}
\title{Holydays to fit prophet hasta 2014}
\usage{
holydays_to_prophet(datos, from, to, ...)
}
\arguments{
\item{filter}{filtro de cuales eventos serán usados, por default "all", para alguna lista en particular usar por ejemplo c("Navidad","BuenFin").}
}
\value{
Data.frame con fechas y rangos de efectos de promociones.
}
\description{
Histórico de fechas donde hay eventos "relevantes" en México para retail. La lista completa se incluye en details.
}
\details{
Los eventos relevantes son:
   \itemize{\item{Navidad:}{ Todo Diciembre centrado al 23 de diciembre.}}
   \itemize{\item{Buen Fin:}{ Varia con los años}}
   \itemize{\item{Black Friday:}{ Tercer jueves de noviembre}}
   \itemize{\item{Cyber Monday:}{ Lunes despues del Black Friday}}
   \itemize{\item{Hot Sale:}{ Varia entre años}}
   \itemize{\item{Independencia:}{ 16 de septiembre MX}}
   \itemize{\item{Halloween:}{ Finales de octubre}}
   \itemize{\item{Pre_Halloween:}{ Compras a inicio de octubre}}
   \itemize{\item{San_valentin:}{ 14 de Febrero}}
   \itemize{\item{caida_25_diciembre:}{ Efecto de "pausa" economica por dia despues a navidad}}
}
\examples{

holydays_to_prophet()
}
