% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{init}
\alias{init}
\title{Rutina Principal de TSRutina}
\usage{
init(datos, frecuencia = NULL, inicio = NULL, init_ = TRUE, ...)
}
\arguments{
\item{datos}{Datos para el analisis de serie de tiempo suport (data.frame and ts class)}

\item{frecuencia}{Frecuencia de la serie de tiempo, sirve para
reescribir la frecuencia cuando datos es un objeto ts}

\item{inicio}{Inicio de la serie de tiempo, igual que frecuencia
sobreescribe valores de objetos ts}

\item{init_}{(True or False) validar el parametro datos}

\item{...}{Not work}
}
\value{
La salida no es como tal un objeto, si no una serie de impresiones de varios
  analisis. La siguiente lista detalla alguno de ellos:
  \itemize{\item{Plost}{  Arroja una lista de plots que ayudan a ver el comportamiento de la serie y como ciertos ajustes se aproximan mejor a ella}}
  \itemize{\item{Resumenes}{  Arroja ciertos resumenes de ciertos ajustes o pruebas que se hacen}}
  \itemize{\item{Modelo}{  Modelo con el menor MSE(Error cuadratico medio)}}
}
\description{
Rutina Principal de TSRutina
}
