% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acf_pacf_CI.R
\name{intervalo_confianza_acf}
\alias{intervalo_confianza_acf}
\title{Obten el intervalo de confianza de un objeto ACF o PACF}
\usage{
intervalo_confianza_acf(
  x,
  ci = 0.95,
  type = "h",
  xlab = "Lag",
  ylab = NULL,
  ylim = NULL,
  main = NULL,
  ci.col = "blue",
  ci.type = c("white", "ma"),
  max.mfrow = 6,
  ask = Npgs > 1 && dev.interactive(),
  mar = if (nser > 2) c(3, 2, 2, 0.8) else par("mar"),
  oma = if (nser > 2) c(1, 1.2, 1, 1) else par("oma"),
  mgp = if (nser > 2) c(1.5, 0.6, 0) else par("mgp"),
  xpd = par("xpd"),
  cex.main = if (nser > 2) 1 else par("cex.main"),
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{x}{Objeto ACF o PACF}

\item{ci}{Valor de Alpha, por defecto 0.95}

\item{type}{Tipo de linea en la grafica (No usar, deshabilitado)}

\item{xlab}{(No usar, deshabilitado)}

\item{ylab}{(No usar, deshabilitado)}

\item{ylim}{(No usar, deshabilitado)}

\item{main}{(No usar, deshabilitado)}

\item{ci.col}{(No usar, deshabilitado)}

\item{ci.type}{Tipo de intervalo de confianza por defecto es white}

\item{...}{Parametros que se pueden pasar a el grafico}
}
\value{
Valor numerico que representa el limite del intervalo de confianza
}
\description{
Atravez del método usado en Stats se obtiene el valor de los intervalos de confianza
}
\details{
Esta pensado para su uso en Rutinas de la paqueteria TSRutina, aunque su uso como función independiente
deberia ser sencillo, se recomienda usarse solo si se conoce sobre el codigo y el tema.
}
\examples{
base=data.frame(x=seq(Sys.Date(),by="days",length=20),y=(rexp(50)+1)*sin(1:50))
p=acf(base$y,plot=FALSE)
intervalo_confianza_acf(p) #0.4382613
}
