% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recomendacion_arma.R
\name{matriz_eacf}
\alias{matriz_eacf}
\title{Matriz extendida de acf}
\usage{
matriz_eacf(z, ar.max = 7, ma.max = 13, print_matrix = TRUE)
}
\arguments{
\item{z}{Serie de tiempo a analizar}

\item{ar.max}{Maximo valor del orden de AR}

\item{ma.max}{Maximo valor del orden de MA}

\item{print_matrix}{Indicador de si se emprime o no la matriz eacf}
}
\value{
Lista de valores del analisis eacf: matriz de acf extendida
   Symbol: matrix con los valores de cuales cordenadas de ARMA son significativos "x" para ese caso, "o" para no significativos.
}
\description{
Se calcula la matriz extendida, igual que en TSA solo que con un control de impresión de matriz
}
\examples{

matriz_eacf(AirPassengers)
}
