% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mensaje.R
\name{mensaje}
\alias{mensaje}
\title{Mensaje con retorno}
\usage{
mensaje(..., domain = NULL, appendLF = TRUE)
}
\arguments{
\item{...}{Conjunto de elementos caracter a mostrar en el mensaje}

\item{domain}{See \code{\link{message}}}

\item{appendLF}{See \code{\link{message}}}
}
\value{
Cadena de texto del mensaje impreso
}
\description{
Similar a message solo que retorna el mensaje.
}
\examples{
a <- mensaje("Hola"," Cristhiandi")
### Aparece impreso "Hola Cristhiandi" en formato message
print(a) ## imprime cadena "Hola Cristhiandi"


}
