% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ajuste_prophet.R
\encoding{UTF-8}
\name{outliers_to_prophet}
\alias{outliers_to_prophet}
\title{Outliers to prophet}
\usage{
outliers_to_prophet(datos, from, to, ..., verbose = T)
}
\arguments{
\item{datos}{Datos para el análisis de serie de tiempo solo Data.frame}

\item{from}{[as.Date] Univalor o vector de fechas para hacer intervalos de outliers}

\item{to}{[as.Date] Univalor o vector de fechas para hacer intervalos de outliers}

\item{...}{Not work}
}
\value{
Data.frame con outliers etiquetados con NA
}
\description{
Modifica los datos que seran outliers, soporta vector de rangos.
Es necesario usar formato de entrada prophet, de no hacerlo se ajustara usando \code{\link{data_to_prophet}}
}
\details{
Si se ingresa vector de fechas, from y to deben ser del mismo tamaño.
}
\examples{

datos <- read.csv("https://raw.githubusercontent.com/crissthiandi/datos/master/Series_tiempo/sunspot_month_dataframe.csv")

outliers_to_prophet(datos,as.Date('2002-02-01'),as.Date('2007-06-01'))

# Si se usan varios rangos
from <- c(as.Date('1800-06-01'),as.Date('2004-03-01'))
to <- c(as.Date('1805-06-01'),as.Date('2007-03-01'))
outliers_to_prophet(datos,from,to)
}
