% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\encoding{UTF-8}
\name{paquetes.tsrutina}
\alias{paquetes.tsrutina}
\title{Carga paquetes para un análisis de series de tiempo}
\usage{
paquetes.tsrutina()
}
\value{
Mensaje si todo sale bien
}
\description{
Ideal para usar en trabajos similar a tidyverse
escribir código para análisis de series de tiempo
}
\examples{

paquetes.tsrutina()
}
