% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{pausa}
\alias{pausa}
\title{Pausa la consola al imprimir resultados}
\usage{
pausa(duracion = Inf)
}
\arguments{
\item{duracion}{Tiempo de pausa en segundos, default Inf (infinito)}
}
\value{

}
\description{
Realiza una pausa tanto indeterminada como por una cantidad de segundos
}
\details{
Cuando \bold{duracion} no es definido la pausa es indefinida y se espera la entrada de "stop" o Esc para terminar la pausa
}
\examples{

#para 10 segundos
pausa(10)

}
