% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{recomendaciones_arma}
\alias{recomendaciones_arma}
\title{Recomendación de modelo ARMA}
\usage{
recomendaciones_arma(time_series, print_matrix = TRUE, msg = FALSE)
}
\arguments{
\item{time_series}{Objeto Serie de tiempo}

\item{print_matrix}{Indicador, imprimir o no matriz de eacf}

\item{msg}{Mostrar mensaje en lugar de vector}
}
\value{
Vector de longitud 2, primera entrada valor de \bold{p}, segunda valor de \bold{q}
}
\description{
Usando una matriz eacf de TSA paqueteria se propone un posible vector con los valores de p y q de un modelos ARMA(p,q)
}
\details{
Se utiliza un metodo de busqueda de esquinas para proponer el valor de
p,q para ARMA(p,q).
}
\examples{
recomendaciones_arma(AirPassengers)
}
