% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{separador}
\alias{separador}
\title{Repide un texto y lo imprime en la consola}
\usage{
separador(repite = "-", num_repetidas = 45, color = "red")
}
\arguments{
\item{repite}{String, que se repetira, idealmente del tipo "-" "_" "*" etc.}

\item{num_repetidas}{Int, Cuantas veces se repite el string anterior.}

\item{color}{Color del texto impresa, soportamos red, blue, green, magenta, silver y yellow}
}
\value{
Impresión en cosola
}
\description{
Repite un texto de forma seriada y se imprime en consola, ideal para hacer una separación entre elementos impresos en consola
}
\examples{

separador("-",10,"blue")
separador()

}
