% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{serie_tiempo_ARIMA}
\alias{serie_tiempo_ARIMA}
\title{Ajuste de Modelo ARIMA a tu Serie de tiempo}
\usage{
serie_tiempo_ARIMA(datos, frecuencia = NULL, inicio = NULL, init_ = FALSE)
}
\arguments{
\item{datos}{Dataframe de no mas de 2 columnas, en el orden primero tiempo y
luego valor el tiempo va en formato fecha, y tiene que ser en el orden
dia-mes-year. La versión 2.1 soporta objetos de la clase TS (time series)}

\item{frecuencia}{Este es el periodo de la serie, trimestral = 3, cuatrimestral = 4
,mensual = 12, etc.}

\item{inicio}{Este es el year a iniciar la serie de tiempo}

\item{init_}{Boleano, True/False indica di se vericaran los datos}
}
\value{
La salida no es como tal un objeto, si no una serie de impresiones de varios
   analisis. El mejor basando en criterio AIC.
}
\description{
Ajusta un modelo arima usando un asistente que te mostrar si tu serie se ajusta a un modelo ARIMA en particular
}
\details{
A diferencia de otras funciones esta utiliza el conocimiento del experto para ajustar el mejor modelo ARIMA a la serie de tiempo estudiada, mientras propone modelos que el asistente considera utiles.
}
\examples{
serie_tiempo_ARIMA(sunspot.year,5)

base=data.frame(tiempo=seq(Sys.Date(),by="days",length=20),valores=1:20*3+runif(20))
serie_tiempo_ARIMA(datos=base,frecuencia=4,inicio=2010)

}
