% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{serie_tiempo_plots}
\alias{serie_tiempo_plots}
\title{Grafica de ajuste de TimeSeries en PNG}
\usage{
serie_tiempo_plots(datos, frecuencia = NULL, inicio = NULL, init_ = FALSE)
}
\arguments{
\item{datos}{Dataframe de no mas de 2 columnas, en el orden primero tiempo y
luego valor el tiempo va en formato fecha, y tiene que ser en el orden
dia-mes-year. La versión 2.1 soporta objetos de la clase TS (time series)}

\item{frecuencia}{Este es el periodo de la serie, trimestral = 3, cuatrimestral = 4
,mensual = 12, etc.}

\item{inicio}{Este es el year a iniciar la serie de tiempo}

\item{init_}{Boleano, True/False indica di se vericaran los datos}
}
\value{
La salida no es como tal un objeto, si no una serie de impresiones de varios
  varios analisis.
  \itemize{\item{Plost}{  Arroja una lista de plots que ayudan a ver el comportamiento de la serie y como ciertos ajustes se aproximan mejor a ella}}
  \itemize{\item{PNG}{ Imagenes grabadas en el directorio de trabajo}}
}
\description{
Obten archivos .png de tus graficos de serie de tiempo en tu directorio
}
\details{
Se usa una salida interactiva para el proceso
}
\examples{
serie_tiempo_rutina(sunspot.year,5)

base=data.frame(tiempo=seq(Sys.Date(),by="days",length=20),valores=1:20*3+runif(1))
serie_tiempo_rutina(datos=base,frecuencia=4,inicio=2010)

}
\author{
Cristhian Diaz
}
