% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\encoding{UTF-8}
\name{serie_tiempo_pruebas}
\alias{serie_tiempo_pruebas}
\title{Pruebas estadísticas a una serie de tiempo}
\usage{
serie_tiempo_pruebas(
  datos,
  frecuencia = NULL,
  validar_ = FALSE,
  msg = TRUE,
  pausa_off = 1
)
}
\arguments{
\item{datos}{Data.frame o objeto TS a analizar}

\item{frecuencia}{Frecuencia de los datos, en caso de TS sobrescribe los valores}

\item{validar_}{Indicador para verificar los datos [True/False]}
}
\value{
Solo arroja la decisión a tomar, por defecto con respecto a p=0.05
}
\description{
Esta función incluye el calculo y la decisión de dos pruebas estadísticas.
}
\details{
Realiza las siguientes pruebas:
   \itemize{\item{Dickey-Fuller:}{ La prueba de estacionariedad}}
   \itemize{\item{Durbin-Watson:}{ La prueba de autocorrelacion}}

Recibe un dataframe de dos columnas, la primera el tiempo y
la segunda el valor de la serie.

Tambien soporta el uso de objetos TimeSeries.
}
\examples{
 base=data.frame(tiempo=seq(Sys.Date(),by="days",length=20),valores=(rexp(50)+1)*sin(1:50))
 serie_tiempo_pruebas(datos=base,frecuencia=4)

}
