% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{tratamiento.fechas.TRS}
\alias{tratamiento.fechas.TRS}
\title{Tratamiento de fechas TSR}
\usage{
tratamiento.fechas.TRS(fecha_vector)
}
\arguments{
\item{fecha_vector}{Vector de valores con las fechas en formato caracter o Date
Formatos soportador:  "\%d/\%m/\%Y", "\%Y-\%m-\%d" y "\%d-\%m-\%Y"}
}
\value{
Vector en formato fecha estandar de R
}
\description{
Tratamiento para que una fecha tenga el formato Date. De no tener un formato convertible se retorna NA
values
}
\examples{
tiempo=seq(Sys.Date(),by="days",length=20)
tratamiento.fechas.TRS(tiempo)

}
