% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{tratamiento.ts_set}
\alias{tratamiento.ts_set}
\title{From ts to data.frame}
\usage{
tratamiento.ts_set(datosts)
}
\arguments{
\item{datosts}{objeto ts a tranformar en data.frame class}
}
\value{
a data.frame object
}
\description{
Convierte un objeto serie de tiempo a data.frame con dos columnas
  x => variable fecha
  y => variable valor
  La variable fecha inicia con el inicio de la serie y termina con el numero de
  frecuencias que se pueden hacer
}
\examples{

tratamiento.ts_set(sunspot.year)

}
