% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{paquetes.tsrutina}
\alias{paquetes.tsrutina}
\title{Carga paquetes para un analisis de series de tiempo}
\usage{
paquetes.tsrutina()
}
\value{
Mensage si todo sale bien
}
\description{
Proximamento en desuso en las funciones de TSRutina pero ideal para usarse en procesos
escribir codigo para analisis de series de tiempo
}
\examples{

paquetes.tsrutina()
}
