% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{recomendacion_autocorrelaciones}
\alias{recomendacion_autocorrelaciones}
\title{Recomendaciones a un grafico ACF o PACF}
\usage{
recomendacion_autocorrelaciones(objeto_cf)
}
\arguments{
\item{objeto_cf}{objeto ACF o PACF}
}
\value{

}
\description{
Obten un recomendación a tu ajuste de series de tiempo ACF o PACF
}
\examples{
base=data.frame(x=seq(Sys.Date(),by="days",length=20),y=1:20*3+runif(1))
recomendacion_autocorrelaciones(acf(base$y,plot = FALSE))

}
