% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{serie_tiempo_plots}
\alias{serie_tiempo_plots}
\title{Grafica de ajuste de TimeSeries en PNG (low computational cost)}
\usage{
serie_tiempo_plots(
  datos,
  frecuencia = NULL,
  inicio = NULL,
  validar_ = FALSE,
  pausa_off = 1
)
}
\arguments{
\item{datos}{Dataframe de no mas de 2 columnas, en el orden primero tiempo y
luego valor el tiempo va en formato fecha, y tiene que ser en el orden
dia-mes-year. La versión 2.1 soporta objetos de la clase TS (time series)}

\item{frecuencia}{Este es el periodo de la serie, trimestral = 3, cuatrimestral = 4
,mensual = 12, etc.}

\item{inicio}{Este es el year a iniciar la serie de tiempo}

\item{validar_}{Boleano, True/False indica si se vericarán los datos}
}
\value{
La salida no es como tal un objeto, si no una serie de impresiones de varios
   analisis.
   \itemize{\item{\bold{Plost}}{  Arroja una lista de plots que ayudan a ver el comportamiento de la serie y como ciertos ajustes se aproximan mejor a ella}}
   \itemize{\item{\bold{PNG}}{ Imagenes grabadas en el directorio de trabajo}}
}
\description{
Obtén archivos .png de tus gráficos de serie de tiempo en tu directorio
}
\details{
Se usa una salida interactiva para el proceso
}
\examples{
serie_tiempo_rutina(sunspot.year,5)

base=data.frame(tiempo=seq(Sys.Date(),by="days",length=20),valores=(rexp(50)+1)*sin(1:50))
serie_tiempo_rutina(datos=base,frecuencia=4,inicio=2010)

}
\author{
Cristhian Diaz
}
