% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\name{serie_tiempo_pruebas}
\alias{serie_tiempo_pruebas}
\title{Pruebas de una serie de tiempo}
\usage{
serie_tiempo_pruebas(
  datos,
  frecuencia = NULL,
  init_ = FALSE,
  msg = TRUE,
  pausa_off = 1
)
}
\arguments{
\item{datos}{Data.frame o objeto TS a analizar}

\item{frecuencia}{Frecuencia de los datos, en caso de TS sobrescribe los valores}

\item{init_}{Indicador para verificar los datos [True/False]}
}
\value{
Solo arroja la decicion a tomar, por defecto con respecto a p=0.05
}
\description{
Esta funcion incluye el calculo y la decision de dos pruebas estadisticas a una serie de tiempo.
}
\details{
Realiza las siguientes pruebas:
La prueba de estacionalidad (Dickey-Fuller)

La prueba de autocorrelacion (Durbin-Watson)

Se debe meter un dataframe de dos columnas, la primera el tiempo y la segunda el valor de la serie
Tambien soporta el uso de objetos TimeSeries
}
\examples{
 base=data.frame(tiempo=seq(Sys.Date(),by="days",length=20),valores=(rexp(50)+1)*sin(1:50))
 serie_tiempo_pruebas(datos=base,frecuencia=4)

}
