% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_de_tiempo.R
\encoding{UTF-8}
\name{tratamiento.ts_set}
\alias{tratamiento.ts_set}
\title{Convert ts to data.frame}
\usage{
tratamiento.ts_set(datosts)
}
\arguments{
\item{datosts}{objeto ts a tranformar en data.frame class}
}
\value{
a data.frame object
}
\description{
Convierte un objeto serie de tiempo a data.frame con dos columnas
  x => variable fecha
  y => variable valor
  La variable fecha es inicializada con el valor start() de la serie de tiempo y termina con el numero de
  frecuencias que se pueden hacer
}
\details{
Si el objeto TS contiene una frecuencia no compatible con las programadas, se te pedirá ingresar el nombre del vector
  (previamente creado en tu enviroment a.k.a entorno de desarrollo) con las fechas a usar. Mira el ejemplo 2.
}
\examples{

tratamiento.ts_set(sunspot.year)

## Ejemplo 2
\dontrun{
  ### Se crea el vector con las fechas
  fechas <- c("2021-05-01","2020-10-05") \%>\% as.Date()
  ### al correr tratamiento.ts_set(datos) se ingresa el nombre : fechas
}


}
